Add multiple preview URL whitelist entries
If you only have a single preview hostname and URL, it should be fully sufficient to set `studio.previewUrlPrefix` to that URL. Due to the CoreMedia issue CMS-21564, the Studio does not correctly handle the case where only `studio.previewUrlPrefix` is set, or `studio.preview-url-whitelist` has only a single entry. This results in error messages in the browser console similar to `Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('https://preview.127.0.0.1.nip.io') does not match the recipient window's origin ('https://studio.127.0.0.1.nip.io').`

Add the preview URL twice to `studio.preview-url-whitelist` to work around this.