Use correct environment variable SOLR_HEAP
In order to specify custom memory management options suitable for running Solr in a container environment, SOLR_HEAP has to be unset or set to the empty string. Previously, the variable JAVA_HEAP was set instead, and Solr used it's default value for SOLR_HEAP, which configures a fixed heap size.

Closes #52