Allow configuring the site mapping protocol
Previously, the site mapping protocol was set to "//" (leading to URLs in the form of "//www.example.com"). While this works in the browser (which will use the protocol of the current page), other clients might be confused by the missing protocol, for example in sitemap.xml or custom resources.

Add `defaults.siteMappingProtocol and the `protocol` property for sitemap entries to be able to adjust the protocol. The default has been changed from "//" to "http://", under the assumption that all public-facing websites are likely to always be https.