Add automountServiceAccountToken to values.yaml
Allow setting automountServiceAccountToken on the pod. The default is ~ (Yaml null), which has the same effect as not specifying it in the pod spec. Kubernetes will take the default from the ServiceAccount in this case.

Note that setting this to false will likely stop the operator from working, unless you can provide some other means for the operator to connect to the Kubernetes API.

See https://kubernetes.io/docs/tasks/configure-pod-container/configure-service-account/#opt-out-of-api-credential-automounting