When configuring Solr followers, create cores on them.
Fixes #59

While there was code to create the cores in the followers, it never ran. Now, there is a specific configuration for the `solr` component, `extra.coresToReplicate` that specifies which cores should be made available for replication on the followers. The `live` core is always on this list implicitly.

While here, docuemnt the `with.handlerPrefixes` a bit better.