Properly handle custom root passwords for MongoDb and MySQL
Previously, the secrets for the database root passwords were hardcoded in the Operator; when overriding the secrets through clientSecretRefs, the Operator would still try to find a secret "mongodb-root" or "mysql-root", respectivly, but fail, since these were not created. Now the configured secrets are used when creating MongoDb and/or MySQL instances.