Allow configuring Ingress response timeouts
The Ingress controller typically waits a specific time for the backend service to start sending a response. Once this timeout is exceeded, an error message is returned. For some applications, it might be necessary to increase that timeout. `with.responseTimeout.live`, `with.responseTimeout.preview`, and `with.responseTimeout.studio` allow specifying the desired timeout in seconds for the respective Ingress resources.